% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/etea_features.R
\name{etea_features}
\alias{etea_features}
\title{Create a list of the Document Frequency Matrix according to minimum word frequency}
\usage{
etea_features(textColumn, stateNum = NULL, timeNum = NULL,
  docvaragg = "null", use_stopwords = TRUE,
  stopwords_language = "english", add_stopwords = NULL,
  remove_stopwords = NULL, verbose = TRUE, toLower = FALSE,
  stem = FALSE, keptFeatures = NULL, removeFeatures = TRUE,
  language = "english", valuetype = c("glob"), thesaurus = NULL,
  dictionary = NULL, removeNumbers = TRUE, removePunct = TRUE,
  removeSeparators = TRUE, removeHyphens = TRUE, removeTwitter = TRUE,
  ngrams = 1L, skip = 0L, concatenator = "_", simplify = FALSE,
  convert_to_tm = TRUE, termNum = 1, ...)
}
\arguments{
\item{textColumn}{character vector containing the text to be analysed; mandatory}

\item{stateNum}{numeric vector containing identifiers for the condition or state when the document or note was recorded/written that it be correctly allocated in the event of more than one note or record being in a state; default NULL}

\item{timeNum}{numeric verctor containing an index that identifies when the document was recorded/noted to give a temporal record.  This normalises progress and case note recording as a progress through a system.  Typically days or minutes after the system commenced; default NULL}

\item{docvaragg}{specifies how the aggregation on docvars is to occur either s stateID only, t timestamp only, st state and timestamp or timestamp and state ; default NULL; Options s, t, st, ts}

\item{use_stopwords}{specifies whether stopwords are to be removed from the corpus (TRUE) or not removed, (FALSE).  Users are reminded that system (language-specific) stopwords may need additions or removals to tailor for a specific need; default TRUE}

\item{add_stopwords}{a character vector of words to be added to the stopwords vector (if any); default is NULL.}

\item{remove_stopwords}{a character vector of words to be removed to the stopwords vector (if any); default is NULL.}

\item{verbose}{to see useful progress information; default is TRUE}

\item{toLower}{to convert all inbound text into lower case. Notably this will degrade the sentence splitting function if applied; default is FALSE; see: \code{\link[quanteda]{tokenize}}}

\item{stem}{reduce word length to root; default is FALSE; see: \code{\link[quanteda]{tokenize}}}

\item{removeFeatures}{remove particular features from inbound text as specified in a list; default is TRUE; see: \code{\link{quanteda}}}

\item{language}{to define local language; default is "english" see: \code{\link{quanteda}}}

\item{valuetype}{to define patterning; default is \code{glob}; see: \code{\link{quanteda}}}

\item{removeNumbers}{remove individual numbers from inbound text, (note: numbers already aggregated with characters such as 1st or 2nd are unaffected); default is TRUE; see: \code{\link{quanteda}}}

\item{removePunct}{remove punctuation from inbound text; default is TRUE; see: \code{\link{quanteda}}}

\item{removeSeparators}{remove separators from inbound text; default is TRUE; see: \code{\link{quanteda}}}

\item{removeHyphens}{remove hyphen characters from inbound text; default is TRUE; see: \code{\link{quanteda}}}

\item{removeTwitter}{remove twitter api characters from inbound text; default is TRUE; see: \code{\link{quanteda}}}

\item{ngrams}{integer vector specifying the number of elements to be concatenated in each ngram; default is 1L; see: \code{\link[quanteda]{ngrams}}}

\item{skip}{integer vector specifying the adjacency skip size for tokens forming the ngrams; \code{0}: see: \code{\link[quanteda]{ngrams}}}

\item{concatenator}{character for combining words, default is \code{_}; see: \code{\link[quanteda]{ngrams}}}

\item{simplify}{character vector of tokens rather than a length of texts; default is FALSE; see: \code{\link[quanteda]{tokenize}}}

\item{convert_to_tm}{logical specifying the requirement for the matrix to be returned in the tm TRUE or quanteda FALSE format}

\item{termNum}{integer specifying the minimum frequency a word is to have been found in the matrix}

\item{...}{Extra arguments, not used}

\item{choice}{of language to determine the content of the basic stopword list; default \code{english}. See \code{\link{quanteda}} for further information.}
}
\value{
a word vector.
}
\description{
This function is a wrapper for a Document Frequency Matrix provided by the function \code{\link[quanteda]{create_q_matrix}} so that it can typically be used to add terms to a lexicon.  The terms to be added to a lexicon must first be categorised in a format that matches the example in the data directory.  Parameters listed below match those of the function \code{\link[quanteda]{create_q_matrix}}.
}
\examples{
## LOAD ##
text_df <- read.csv("data/militant_suffragette_extract.csv",header=TRUE, sep=";")
textColumn<-as.character(text_df$Notes) # typically textual interview or clinical notes
## CREATE FEATURES LIST ##
features_vec <- etea_features(textColumn, termNum=1,verbose=TRUE,use_stopwords=TRUE)
features_vec

}
\author{
Chris Kirk
}

